requirements.txt
addict==2.1.0
click==6.7
Django==1.11
djangorestframework==3.6.2
drf-autodocs==0.4.3
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
mysqlclient==1.3.10
olefile==0.44
openapi-codec==1.3.1
Pillow==4.1.0
Pygments==2.2.0
pytz==2017.2
requests==2.13.0
simplejson==3.10.0
uritemplate==3.0.0
